+++ contents/ui/configAppearance.qml
@@ -46,6 +46,8 @@ KCM.SimpleKCM {
     property var cfg_partitionOptOutDefault
     property var cfg_partitionLabels
     property var cfg_partitionLabelsDefault
+    property var cfg_diskPartitionColors
+    property var cfg_diskPartitionColorsDefault
     property var cfg_showCpuCores
     property var cfg_showCpuCoresDefault
     property var cfg_mergeCpuTemp